+++ M15_JocDeDaus/src/main/java/com/ITAcademy/M15_JocDeDaus/Services/IGameService.java
@@ -10,7 +10,7 @@ public interface IGameService {
 	public GameDTO saveGame(String player_id);
 	public List<GameDTO> gamesByPlayer(String player_id);
 	public void deleteGamesByPlayer(String player_id);
-	public BigDecimal calculateWinRate(String player_id);
+	public double calculateWinRate(String player_id);
 	public Boolean checkNoGames();
